docs: move common stuff back into barchart, document design choices in readme
